requirements.txt
torch
Pillow
transformers
gradio
